custom rng
for some reason I decided to invest a good dozens of hours on implementing a custom random number generator.

It's not particularly well done structure-wise, but it's built in a way that if I want to add a new rng algorithm, I can just create a new .cs file in the Generators folder, code in the algorithm and I don't need to worry about anything else.